get explore and annotate components
